bug fixed with triple aces
since we are only allowing one split, there is a bug where there can be a soft 12 even though basic strategy does not account for this. Added a line for soft 12 in basic strategy which is the same as soft 13. This error only occurs when aces are split and another ace is dealt.